% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_leading_symbols.R
\name{add_leading_symbols}
\alias{add_leading_symbols}
\title{Create indices padded with zeros}
\usage{
add_leading_symbols(file_names)
}
\arguments{
\item{file_names}{vector of any values.}
}
\value{
A string with numbers padded with leadinng zero.
}
\description{
Create indices padded with zeros. This is important for creating appropriate
for sorting names.
}
\examples{
add_leading_symbols(1:200)

}
\author{
George Moroz <agricolamz@gmail.com>
}
