% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/annotate_textgrid.R
\name{annotate_textgrid}
\alias{annotate_textgrid}
\title{Annotate textgrid}
\usage{
annotate_textgrid(annotation, textgrid, tier = 1, each = 1,
  backup = TRUE)
}
\arguments{
\item{annotation}{vector of stimuli (obligatory)}

\item{textgrid}{character with a filename or path to the TextGrid}

\item{tier}{value that could be either ordinal number of the tier either name of the tier}

\item{each}{non-negative integer. Each element of x is repeated each times}

\item{backup}{logical. If TRUE (by default) it creates a backup tier.}
}
\description{
Annotates textgrids. It is possible to difine step in the argument "each", so each second element of the tier will be annotated.
}
\examples{
# annotate_textgrid(annotation = c("a", "b"), textgrid = "exampl.TextGrid")

}
\author{
George Moroz <agricolamz@gmail.com>
}
