% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/audacity_to_df.R
\name{audacity_to_df}
\alias{audacity_to_df}
\title{Audacity's labels to dataframe}
\usage{
audacity_to_df(file_name)
}
\arguments{
\item{file_name}{file_name string with a filename or path to the .txt file produced by Audacity}
}
\value{
a dataframe with columns:  \code{content}, \code{time_start}, \code{time_end}.
}
\description{
Audacity make it possible to annotate sound files with labels that can be exported as a .tsv file with .txt extension. This function convert result to dataframe.
}
\author{
George Moroz <agricolamz@gmail.com>
}
