% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/concatenate_soundfiles.R
\name{concatenate_soundfiles}
\alias{concatenate_soundfiles}
\title{Concatenate sounds}
\usage{
concatenate_soundfiles(path = getwd(), filename, textgrid = TRUE)
}
\arguments{
\item{path}{path to the directory with soundfiles.}

\item{filename}{name of the result files (and TextGrid, if this argument equal TRUE)}

\item{textgrid}{logic value, whether create Praat TextGrid whose interval labels are the original names of the sound}
}
\description{
Creates a sound file and a Praat TextGrid whose interval labels are the original names of the sound
}
\author{
George Moroz <agricolamz@gmail.com>
}
