% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_empty_textgrid.R
\name{create_empty_textgrid}
\alias{create_empty_textgrid}
\title{Create an empty TextGrid}
\usage{
create_empty_textgrid(file_name, tiers = NULL, point_tiers = NULL)
}
\arguments{
\item{file_name}{a reference sound file.}

\item{tiers}{a vector that contain tier names.}

\item{point_tiers}{a vector that defines which tiers should be made point tiers. This argument excepts numeric values (e. g. \code{c(2, 4)} means second and forth tiers) or character (e. g. \code{c("a", "b")} means tiers with names "a" and "b")}
}
\value{
The function returns no output, just creates a Praat TextGrid in the same folder as a reference sound file.
}
\description{
Creates an empty Praat TextGrid in the same folder as a reference sound file. It is possible to manage with predefined number of tiers, their names and their types.
}
\author{
George Moroz <agricolamz@gmail.com>
}
