% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_glossed_document.R
\name{create_glossed_document}
\alias{create_glossed_document}
\title{Create a glossed document}
\usage{
create_glossed_document(
  flextext = NULL,
  output_dir,
  output_file = "glossed_document",
  output_format = "html"
)
}
\arguments{
\item{flextext}{path to a .flextext file.}

\item{output_dir}{the output directory for the rendered file}

\item{output_file}{the name of the result .html file (by default stimuli_viewer)}

\item{output_format}{The option can be "html" or "docx"}
}
\value{
If \code{render} is \code{FALSE}, the function returns a path to the temporary file with .csv file. If \code{render} is \code{TRUE}, there is no output in a function.
}
\description{
Creates a file with glossed example (export from .flextext or other formats)
}
\author{
George Moroz <agricolamz@gmail.com>
}
