% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_image_look_up.R
\name{create_image_look_up}
\alias{create_image_look_up}
\title{Create image look_up objects for html viewer}
\usage{
create_image_look_up(img_src, img_caption = "", text = "&#x1f441;")
}
\arguments{
\item{img_src}{string or vector of strings with a image(s) path(s).}

\item{img_caption}{string or vector of strings that will be displayed when image is clicked.}

\item{text}{string o vector of strings that will be displayed as view link. By default it is eye emoji (&#x1f441;).}
}
\value{
a string or vector of strings
}
\description{
Create image look_up objects for html viewer
}
\examples{
create_image_look_up("path/to/your/file")

}
\author{
George Moroz <agricolamz@gmail.com>
}
