% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_subannotation.R
\name{create_subannotation}
\alias{create_subannotation}
\title{Create boundaries in a texgrid tier}
\usage{
create_subannotation(
  textgrid,
  tier = 1,
  new_tier_name = "",
  n_of_annotations = 4,
  each = 1,
  omit_blank = TRUE,
  overwrite = TRUE
)
}
\arguments{
\item{textgrid}{character with a filename or path to the TextGrid}

\item{tier}{value that could be either ordinal number of the tier either name
of the tier}

\item{new_tier_name}{a name of a new created tier}

\item{n_of_annotations}{number of new annotations per annotation to create}

\item{each}{non-negative integer. Each new blank annotation is repeated every
first, second or ... times}

\item{omit_blank}{logical. If TRUE (by dafault) it doesn't create
subannotation for empy annotations.}

\item{overwrite}{logical. If TRUE (by dafault) it overwrites an existing
tier.}
}
\value{
a string that contain TextGrid. If argument write is \code{TRUE},
then no output.
}
\description{
Create boundaries in a texgrid tier
}
\examples{
create_subannotation(system.file("extdata", "test.TextGrid",
  package = "phonfieldwork"
),
tier = 1, overwrite = FALSE
)
}
\author{
George Moroz <agricolamz@gmail.com>
}
