% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/crocodile.R
\name{crocodile}
\alias{crocodile}
\title{create crocodile's roar}
\usage{
crocodile(iterations)
}
\arguments{
\item{iterations}{integer numeric value}
}
\value{
a string with cute crocodile's roar
}
\description{
Creates a crocodile's roar for a given number of times.
}
\examples{
crocodile(3)

}
\author{
Kudrjashov Sergej <xenomirant@gmail.com>
}
