% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/df_to_exb.R
\name{df_to_exb}
\alias{df_to_exb}
\title{df_to_exb}
\usage{
df_to_exb(
  df,
  name,
  outputPath,
  referenced_file = "",
  ud_meta = NULL,
  speaker_table = NULL
)
}
\arguments{
\item{df}{dataframe that must include columns tier, id, content, tier_name, time_start, time_end and may include columns tier_type, tier_category, tier_speaker}

\item{name}{transcription name}

\item{outputPath}{path}

\item{referenced_file}{name of audio file}

\item{ud_meta}{vector c(parameter:value) of metadata}

\item{speaker_table}{dataframe with speaker data; must include column id with speaker id; may include columns abbreviation, sex and any amount of feature columns (e.g. name, surname)}
}
\value{
file with transcription
}
\description{
Takes a dataframe and turns it into a .exb file
}
\examples{
df_to_exb(Beckhams, 
          name='Beckhams', 
          outputPath = 'beck.xml', 
          referenced_file = 'beck.wav', 
          ud_meta = c('language' = 'eng', 'place' = 'New York'),
          speaker_table=data.frame(id = c('SPK0', 'SPK1', 'SPK2'), 
                                   abbreviation = c('PAR', 'VIC', 'DAV'), 
                                   sex = c('m', 'f', 'f'),
                                   name = c('Michael', 'Victoria', 'David'), 
                                   surname = c('Parkinson', 'Beckham', 'Beckham')))

}
\author{
Valeria Buntiakova <valleriabun@gmail.com>
}
