% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/df_to_tier.R
\name{df_to_tier}
\alias{df_to_tier}
\title{Dataframe to TextGrid's tier}
\usage{
df_to_tier(df, textgrid, name = "")
}
\arguments{
\item{df}{an R dataframe object that contains columns named "annotation", "start" and "end" (if you want an interval tier)}

\item{textgrid}{character with a filename or path to the TextGrid}

\item{name}{vector that contain a name for a created tier}
}
\description{
Convert a dataframe to a Praat TextGrid.
}
\examples{
# tier_to_df("my.TextGrid")

}
\author{
George Moroz <agricolamz@gmail.com>
}
