% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/draw_sound.R
\name{draw_sound}
\alias{draw_sound}
\title{Draw Oscilogram and Spectrogram}
\usage{
draw_sound(path, colores = FALSE, maxfreq = 5000, dynamicrange = 50)
}
\arguments{
\item{path}{a path to the soundfile}

\item{colores}{If TRUE, a color spectrogram will be displayed. If FALSE, greyscale is used. If a vector of colors is provided, these colors are used to create the spectrogram.}

\item{maxfreq}{he maximum frequency to be displayed for the spectrogram up to a maximum of fs/2. This is set to 5000 Hz by default}

\item{dynamicrange}{Values greater than this many dB below the maximum will be displayed in the same color}
}
\value{
Oscilogram and Spectrogram plot.
}
\description{
Create oscilogram and spectrogram plot. Based on \code{phonTools::spectrogram}.
}
\author{
George Moroz <agricolamz@gmail.com>
}
