% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eaf_to_df.R
\name{eaf_to_df}
\alias{eaf_to_df}
\title{ELAN's .eaf file to dataframe}
\usage{
eaf_to_df(file_name, eafs_from_folder = NULL)
}
\arguments{
\item{file_name}{string with a filename or path to the .eaf file}

\item{eafs_from_folder}{path to a folder with multiple .eaf files. If this
argument is not \code{NULL}, then the function goes through all files and
create a merged dataframe for all of them.}
}
\value{
a dataframe with columns:  \code{tier}, \code{id}, \code{content},
\code{tier_name}, \code{tier_type}, \code{time_start}, \code{time_end},
\code{source}).
}
\description{
Convert .eaf file from ELAN to a dataframe.
}
\examples{
eaf_to_df(system.file("extdata", "test.eaf", package = "phonfieldwork"))
}
\author{
George Moroz <agricolamz@gmail.com>
}
