% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/example_textgrid.R
\docType{data}
\name{example_textgrid}
\alias{example_textgrid}
\title{Praat TextGrid example file}
\format{A character vector of the length 52}
\usage{
example_textgrid
}
\description{
Praat TextGrid example file that has two tiers: an interval tier with 5 intervals and a point tier with 4 points.
}
\keyword{datasets}
