% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/example_tg_df.R
\docType{data}
\name{example_tg_df}
\alias{example_tg_df}
\title{Praat TextGrid tier dataframe example file}
\format{A data frame with 5 rows and 4 variables:
\describe{
  \item{id}{an index of each row}
  \item{start}{a start time of the interval}
  \item{end}{an end time of the interval}
  \item{annotation}{an annotation of the interval}
}}
\usage{
example_tg_df
}
\description{
Praat TextGrid interval tier stored as a dataframe. This is an annotation for test.wav (could be reached via \code{system.file("extdata", "test.wav", package = "phonfieldwork")})
}
\keyword{datasets}
