% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_intervals.R
\name{extract_intervals}
\alias{extract_intervals}
\title{Concatenate sounds}
\usage{
extract_intervals(filename, textgrid, tier = 1, prefix = NULL,
  suffix = NULL, autonumber = TRUE, path = getwd())
}
\arguments{
\item{filename}{name of the result files (and TextGrid, if this argument equal TRUE)}

\item{textgrid}{logic value, whether create Praat TextGrid whose interval labels are the original names of the sound}

\item{tier}{tier number or name that should be used as base for extraction and names}

\item{prefix}{character vector containing prefix(es) for file names}

\item{suffix}{character vector containing suffix(es) for file names}

\item{autonumber}{if TRUE automatically add number of extracted sound to the filename. Prevents from creating a duplicated files}

\item{path}{path to the directory with soundfiles.}
}
\description{
Creates a sound file and a Praat TextGrid whose interval labels are the original names of the sound
}
\author{
George Moroz <agricolamz@gmail.com>
}
