% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/flextext_to_df.R
\name{flextext_to_df}
\alias{flextext_to_df}
\title{FLEX's .flextext file to dataframe}
\usage{
flextext_to_df(flextext)
}
\arguments{
\item{flextext}{string with a filename or path to the .flextext file}
}
\value{
a dataframe with columns: \code{txt}, \code{cf}, \code{hn}, \code{gls}, \code{msa}, \code{morph}, \code{word}, \code{phrase}, \code{paragraph}, \code{text}, \code{text_title}
}
\description{
Convert .flextext file from FLEX to a dataframe.
}
\author{
George Moroz <agricolamz@gmail.com>
}
