% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/frog.R
\name{frog}
\alias{frog}
\title{frog}
\usage{
frog(n)
}
\arguments{
\item{length}{of string}
}
\value{
a string of 'kva'
}
\description{
Returns a string of 'kva'
}
\examples{
frog(3)

}
\author{
Valeria Buntiakova <valleriabun@gmail.com>
}
