% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_textgrid_names.R
\name{get_textgrid_names}
\alias{get_textgrid_names}
\title{Extract TextGrid names}
\usage{
get_textgrid_names(textgrid)
}
\arguments{
\item{textgrid}{path to the TextGrid}
}
\value{
return a vector of tier names from given TextGrid
}
\description{
Extract TextGrid names.
}
\examples{
get_textgrid_names(example_textgrid)

}
\author{
George Moroz <agricolamz@gmail.com>
}
