% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_from_folder.R
\name{read_from_folder}
\alias{read_from_folder}
\title{Read multiple files from the folder}
\usage{
read_from_folder(path, FUN)
}
\arguments{
\item{path}{to a folder with multiple sound files.}

\item{FUN}{should be one of the following functions: \code{\link{get_sound_duration}}, \code{\link{audacity_to_df}}, \code{\link{eaf_to_df}}, \code{\link{exb_to_df}}, \code{\link{flextext_to_df}}, \code{\link{formant_to_df}}, \code{\link{intensity_to_df}}, \code{\link{pitch_to_df}}, \code{\link{srt_to_df}}, \code{\link{textgrid_to_df}}}
}
\description{
This function reads multiple files from the folder. The first argument is the path, the second argument is the phonfieldwork function that define the type of files.
}
\examples{

read_from_folder(system.file("extdata", package = "phonfieldwork"), get_sound_duration)

}
\author{
George Moroz <agricolamz@gmail.com>
}
