% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/remove_textgrid_tier.R
\name{remove_textgrid_tier}
\alias{remove_textgrid_tier}
\title{Remove tier from texgrid}
\usage{
remove_textgrid_tier(textgrid, tier, overwrite = TRUE)
}
\arguments{
\item{textgrid}{character with a filename or path to the TextGrid}

\item{tier}{value that could be either ordinal number of the tier either name
of the tier}

\item{overwrite}{logical. If TRUE (by dafault) it overwrites an existing
tier.}
}
\value{
a string that contain TextGrid. If argument write is \code{TRUE},
then no output.
}
\description{
Remove tier from texgrid
}
