% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/set_textgrid_names.R
\name{set_textgrid_names}
\alias{set_textgrid_names}
\title{Rewrite TextGrid names}
\usage{
set_textgrid_names(textgrid, tiers, names, write = TRUE)
}
\arguments{
\item{textgrid}{path to the TextGrid}

\item{tiers}{integer vector with the number of tiers that should be named}

\item{names}{vector of strings with new names for TextGrid tiers}

\item{write}{logical. If TRUE (by dafault) it overwrites an existing tier}
}
\value{
a string that contain TextGrid. If argument write is \code{TRUE},
then no output.
}
\description{
Rewrite TextGrid names.
}
\examples{
set_textgrid_names(system.file("extdata", "test.TextGrid",
                               package = "phonfieldwork"),
tiers = 3, names = "new_name", write = FALSE)

}
\author{
George Moroz <agricolamz@gmail.com>
}
