% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/srt_to_df.R
\name{srt_to_df}
\alias{srt_to_df}
\title{Subtitles .srt file to dataframe}
\usage{
srt_to_df(file_name)
}
\arguments{
\item{file_name}{string with a filename or path to the .srt file}
}
\value{
a dataframe with columns:  \code{id}, \code{content},
\code{time_start}, \code{time_end}, \code{source}.
}
\description{
Convert subtitles .srt file to a dataframe.
}
\examples{
srt_to_df(system.file("extdata", "test.srt", package = "phonfieldwork"))
}
\author{
George Moroz <agricolamz@gmail.com>
}
