% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tier_to_df.R
\name{tier_to_df}
\alias{tier_to_df}
\title{TextGrid's tier to dataframe}
\usage{
tier_to_df(file_name, tier = 1, encoding = "unknown")
}
\arguments{
\item{file_name}{string with a filename or path to the TextGrid}

\item{tier}{value that could be either ordinal number of the tier either name of the tier. By default is '1'.}

\item{encoding}{TextGrid encoding. Import from \code{readLines()} function.}
}
\value{
a dataframe with columns:  \code{id}, \code{time_start}, \code{time_end}, \code{content}
}
\description{
Convert selected tier from a Praat TextGrid to a dataframe.
}
\examples{
tier_to_df(system.file("extdata", "test.TextGrid", package = "phonfieldwork"))

}
\author{
George Moroz <agricolamz@gmail.com>
}
