% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_presentation.R
\name{create_presentation}
\alias{create_presentation}
\title{Create a presentation}
\usage{
create_presentation(stimuli, translations = "", output = "html",
  filename = "stimuli_presentation")
}
\arguments{
\item{stimuli}{vector of stimuli (obligatory)}

\item{translations}{vector of translations (optional)}

\item{output}{string that difine the output type: "html" (by default) or "pptx"}

\item{filename}{name of the result presentation file}
}
\description{
Create a html or powerpoint presentation from list of words and translations.
}
\examples{
## Not run:
create_presentation("rzeka", "river")

}
\author{
George Moroz <agricolamz@gmail.com>
}
