% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.r
\name{create_sound_play}
\alias{create_sound_play}
\title{Create audio play objects for html viewer}
\usage{
create_sound_play(snd_src, text = "&#x1f442;")
}
\arguments{
\item{snd_src}{string or vector of strings with a image(s) path(s).}

\item{text}{string o vector of strings that will be displayed as view link.
By default it is ear emoji (&#x1f442;).}
}
\value{
a string or vector of strings
}
\description{
Create audio play objects for html viewer
}
\author{
George Moroz <agricolamz@gmail.com>
}
