% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/crocodile.R
\name{crocodile}
\alias{crocodile}
\title{creates crocodile's roar}
\usage{
crocodile(iterations)
}
\arguments{
\item{iterations}{integer numeric value}
}
\value{
a string with crocodile's roar
}
\description{
Creates a crocodile's roar for a given number of times.
}
\author{
Kudrjashov Sergej
}
