% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/crocodile.R
\name{crocodile}
\alias{crocodile}
\title{Crocodile's roar}
\usage{
crocodile(iterations)
}
\arguments{
\item{numeric}{value}
}
\value{
a string with crocodile's roar
}
\description{
Creates a crocodile's roar for a given number of times.
}
\examples{
crocodile(4)

}
\author{
Kudrjashov Sergej
}
