% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/draw_sound.R
\name{draw_sound}
\alias{draw_sound}
\title{Draw Oscilogram and Spectrogram}
\usage{
draw_sound(file_name, output_file = NULL, title = NULL,
  spectrum_colors = FALSE, maximum_frequency = 5000,
  dynamic_range = 50, window_length = 5, output_width = 750,
  output_height = 500, output_units = "px",
  sounds_from_folder = NULL, pic_folder_name = "pics", prefix = NULL,
  suffix = NULL, autonumber = FALSE)
}
\arguments{
\item{file_name}{a soundfile}

\item{output_file}{the name of the output file}

\item{title}{the title for the plot}

\item{spectrum_colors}{if TRUE, a color spectrogram will be displayed. If FALSE, greyscale is used. If a vector of colors is provided, these colors are used to create the spectrogram.}

\item{maximum_frequency}{the maximum frequency to be displayed for the spectrogram up to a maximum of fs/2. This is set to 5000 Hz by default}

\item{dynamic_range}{values greater than this many dB below the maximum will be displayed in the same color}

\item{window_length}{the desired analysis window length in milliseconds.}

\item{output_width}{the width of the device}

\item{output_height}{the height of the device}

\item{output_units}{the units in which height and width are given. Can be "px" (pixels, the default), "in" (inches), "cm" or "mm".}
}
\value{
Oscilogram and Spectrogram plot.
}
\description{
Create oscilogram and spectrogram plot. Based on \code{phonTools::spectrogram}.
}
\author{
George Moroz <agricolamz@gmail.com>
}
