% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eaf_to_df.R
\name{eaf_to_df}
\alias{eaf_to_df}
\title{ELAN's .eaf file to dataframe}
\usage{
eaf_to_df(eaf)
}
\arguments{
\item{eaf}{string with a filename or path to the TextGrid}
}
\value{
a dataframe with columns:  \code{tier}, \code{id}, \code{content}, \code{tier_name}, \code{tier_type}, \code{ts_start}, \code{ts_end}, \code{a_id}, \code{ar}).
}
\description{
Convert .eaf file from ELAN to a dataframe.
}
\examples{
# eaf_to_df(example_eaf)

}
\author{
George Moroz <agricolamz@gmail.com>
}
