% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eaf_to_df.R
\name{eaf_to_df}
\alias{eaf_to_df}
\title{ELAN's .eaf file to dataframe}
\usage{
eaf_to_df(file_name)
}
\arguments{
\item{file_name}{string with a filename or path to the .eaf file}
}
\value{
a dataframe with columns:  \code{tier}, \code{id}, \code{content},
\code{tier_name}, \code{tier_type}, \code{tier_ref}, \code{event_local_id},
 \code{dependent_on}, \code{time_start}, \code{time_end}, \code{source}, \code{media_url}
and attributes: \code{MEDIA_URL}, \code{MIME_TYPE}, \code{RELATIVE_MEDIA_URL}.
}
\description{
Convert .eaf file from ELAN to a dataframe.
}
\examples{
eaf_to_df(system.file("extdata", "test.eaf", package = "phonfieldwork"))
}
\author{
George Moroz <agricolamz@gmail.com>

Kudrjashov Sergej <xenomirant@gmail.com>
}
