% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_sound_durations.R
\name{get_sound_duration}
\alias{get_sound_duration}
\title{Get file(s) duration}
\usage{
get_sound_duration(file_name)
}
\arguments{
\item{file_name}{a sound file}
}
\description{
Calculate sound(s) duration.
}
\examples{
get_sound_duration(
  system.file("extdata", "test.wav", package = "phonfieldwork")
)
}
\author{
George Moroz <agricolamz@gmail.com>
}
