% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_sound_durations.R
\name{get_sound_duration}
\alias{get_sound_duration}
\title{Get file(s) duration}
\usage{
get_sound_duration(file_name, sounds_from_folder = NULL)
}
\arguments{
\item{file_name}{a sound file}

\item{sounds_from_folder}{path to a folder with multiple sound files. If this argument is not \code{NULL}, then the function goes through all files and calculates duration for all of them.}
}
\description{
Calculate sound(s) duration.
}
\examples{
get_sound_duration(
  sounds_from_folder = system.file("extdata", package = "phonfieldwork"))

}
\author{
George Moroz <agricolamz@gmail.com>
}
