% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_textgrid_names.R
\name{get_textgrid_names}
\alias{get_textgrid_names}
\title{Extract TextGrid names}
\usage{
get_textgrid_names(textgrid, encoding = "unknown")
}
\arguments{
\item{textgrid}{path to the TextGrid}

\item{encoding}{TextGrid encoding. Import from \code{readLines()} function.}
}
\value{
return a vector of tier names from given TextGrid
}
\description{
Extract TextGrid names.
}
\examples{
get_textgrid_names(system.file("extdata", "test.TextGrid", package = "phonfieldwork"))

}
\author{
George Moroz <agricolamz@gmail.com>
}
