% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/textgrid_to_df.R
\name{textgrid_to_df}
\alias{textgrid_to_df}
\title{TextGrid to dataframe}
\usage{
textgrid_to_df(textgrid, encoding = "unknown")
}
\arguments{
\item{textgrid}{string with a filename or path to the TextGrid}

\item{encoding}{TextGrid encoding. Import from \code{readLines()} function.}
}
\value{
a dataframe with columns:  \code{id}, \code{start}, \code{end} (if it is an interval tier -- the same as the start value), \code{annotation}, and \code{tier}
}
\description{
Convert Praat TextGrid to a dataframe.
}
\examples{
textgrid_to_df(example_textgrid)

}
\author{
George Moroz <agricolamz@gmail.com>
}
