% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tier_to_df.R
\name{tier_to_df}
\alias{tier_to_df}
\title{TextGrid's tier to dataframe}
\usage{
tier_to_df(textgrid, tier = 1)
}
\arguments{
\item{textgrid}{string with a filename or path to the TextGrid}

\item{tier}{value that could be either ordinal number of the tier either name of the tier. By default is '1'.}
}
\value{
a dataframe with columns:  \code{id}, \code{start} and \code{end annotation} (if it is an interval tier)
}
\description{
Convert selected tier from a Praat TextGrid to a dataframe.
}
\examples{
tier_to_df(example_textgrid)

}
\author{
George Moroz <agricolamz@gmail.com>
}
